Automatically update java client (ref: Cosmo-Tech/cosmotech-api@762f5d5b11c6a0e7fc4d149fa028a5474653521b)
===
Revert Revert Revert test for True value for ingress as it is set by ARM

Either we are consistent with how we handle such checks everywhere in
the script, or we enforce clients to send the required format.

The ARM Templates have been updated to send a lower-case value, so this
change is a bit redundant.

This reverts commit 0ef8d77f307a3e19bc17bc5b288892b05ada0959.
===

Co-authored-by: rm3l <rm3l@users.noreply.github.com>