Automatically update java client (ref: Cosmo-Tech/cosmotech-api@ecb1a83360705ad6720312ce33a1d512ca6ec0cc)
===
Reduce time to deployments by dropping the calls to 'helm repo update'

'helm repo update' is useful for updating local caches for *all* repos,
in preparation for running 'helm search'.
This is useless when running the deployment script and
may take a lot of time when we have a lot of repos.
===

Co-authored-by: rm3l <rm3l@users.noreply.github.com>