Automatically update java client (ref: Cosmo-Tech/cosmotech-api@9864a4513f8b77cc6bb30011aba08a0389a43e74)
===
chore(CI): Lint Helm Charts with KubeLinter

KubeLinter perform more checks as opposed to 'helm lint'.

For now, those checks will not prevent the build from passing,
but this may change in the near future.

chore(CI): Upload KubeLinter SARIF reports to GitHub Code Scanning

The SARIF format is not supported by the GitHub action.
Therefore, this requires installing and using the kube-linter CLI.

For now, those checks will not prevent the build from passing,
but this may change in the near future.
===

Co-authored-by: rm3l <rm3l@users.noreply.github.com>