Automatically update java client (ref: Cosmo-Tech/cosmotech-api@b9637b4e88fafab66218161e4db9d991c1a0c4bf)
===
chore(CD): Pass NGINX and CertManager env vars to the deployment script

This way, nginx-ingress and cert-manager Helm releases will also be handled here.

chore(CD): Use a repository_dispatch event to trigger the 'deploy' workflow

The 'workflow_run' event does not support passing extra information,
like the exact SHA of the origin Workflow. [1]
This is an issue because if other changes were pushed before the 'deploy'
workflow starts, the 'checkout' Action would pull the latest HEAD,
and we might not deploy the right changes.
Worse, the 'deploy' Workflow might not pass if
the container images and Helm charts were not yet pushed.

[1] https://github.community/t/pass-data-from-original-workflow-that-triggered-current-workflow/148988
===

Co-authored-by: rm3l <rm3l@users.noreply.github.com>